fix some bug for packer
